pyproject.toml
[tool.poetry]
name = "blog"
version = "0.0.1"
description = ""
authors = ["dadevel <dadevel@disroot.org>"]
license = "MIT"
readme = "README.md"
packages = [{include = "build.py"}]

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.3"
markdown = "^3.5.2"
pygments = "^2.17.2"
pymdown-extensions = "^10.7.1"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
